[apptags] Added interval configuration
I noticed that the push_times (line 50) in the collector were nopt
configurable from datadog.conf (because never seeked in config.py). I
fixed that, added an interval config for the agent and added comments
and examples about the in the datadog.conf.example file.

I also removed a piece of code I used to test that the custom interval
config was working.